Added 'imagePath' and changed 'MarkerIcon' image path.
Within the function LTreering, I added the option 'imagePath'. Then changed the path which image icons are displayed. The path is constructed so it will connect through the web to the  'leaflet-treering' project hosted on the server. A path to connect 'index.html' to the server needs to be developped.