Made annotations draggable. Issue #34
Annotations can be dragged and their positions will be saved in the JSON file.

Changed map asset in index.html so marker data would align with core image.

Tested successfully on Firefox and Chrome. Save a local copy does not work on edge so it could not be fully tested.